Move utils to reduce core's length
